+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.4.0](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/tree/1.4.0)
+- Adds `showWeekends` flag in month view to hide & show weekends view. 
+  Default is `showWeekends = true` shows all weekdays. [#385](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/385)
+- Fixes issue: Event is shown when days not in current month is hidden `hideDaysNotInMonth = true`.
+  Events are now hidden for days not in the current month if `hideDaysNotInMonth = true`.
+- Fixes `startHour` and `endHour` not updating when rebuilding in week view. [#410](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/410)
+
 # [1.3.0 - 12 Nov 2024](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/tree/1.3.0)
 
 - Fixes full day event position when fullHeaderTitle is empty.